+++ examples/reactnavigation/src/AppNavigator.js
