+++ examples/reactnavigation/src/__tests__/AppNavigator.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { render, fireEvent, cleanup } from 'react-native-testing-library';
 
-import Navigation from '../Navigation';
+import AppNavigator from '../AppNavigator';
 
 // Resolves the error in https://github.com/facebook/react-native/issues/11094#issuecomment-263240420
 jest.mock('react-native/Libraries/Animated/src/NativeAnimatedHelper');
@@ -13,7 +13,7 @@ describe('Testing react navigation', () => {
   test('page contains the header and 10 items', () => {
     const component = (
       <NavigationContainer>
-        <Navigation />
+        <AppNavigator />
       </NavigationContainer>
     );
 
@@ -29,7 +29,7 @@ describe('Testing react navigation', () => {
   test('clicking on one item takes you to the details screen', async () => {
     const component = (
       <NavigationContainer>
-        <Navigation />
+        <AppNavigator />
       </NavigationContainer>
     );
